requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic_settings
google-genai
httpx
aiogram
gunicorn
aiohttp_socks
psycopg2-binary
python-multipart
httpx[socks]
socksio>=1.0.0
beautifulsoup4
requests
lxml
